Pim6d: Mroute file duplication removal
Mroute duplicate functions removed and made us hybrid for v4 and v6
